+++ R/annotations.R
@@ -264,7 +264,7 @@ mcns_annotate_body <- function(x, test=TRUE, version=NULL, write_empty_fields=FA
 
 # Clio/Dvid DB schema endpoint
 .url_clio_schema = function() {
-  paste0(servers4dataset('CNS')$dvid,
+  paste0(malevnc:::flyem_servers4dataset('CNS')$dvid,
          "/api/node/:master/segmentation_annotations/json_schema")
 }